wip refactor env var types
